+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - Add `salmon.s3/upload!` for uploading to a bucket.
   - Add `com.cognitect.aws/s3` as a dependency.
 - (breaking) Add required argument `:regions` to `salmon.cleanup/delete-all!`.
+- Add `salmon.resource.certificate/dns-validated`. This creates a
+  resource that defines an automatically-validated AWS Certificate Manager
+  (ACM) certificate.
 
 ## v0.12.0 (2023-11-01)